Apply suggestion from @swoboda1337

+++ esphome/components/dlms_meter/dlms_meter.cpp
@@ -28,7 +28,7 @@ void DlmsMeterComponent::dump_config() {
 
 void DlmsMeterComponent::loop() {
   // Read while data is available, netznoe uses two frames so allow 2x max frame length
-  int avail = this->available();
+  size_t avail = this->available();
   if (avail > 0) {
     size_t remaining = MBUS_MAX_FRAME_LENGTH * 2 - this->receive_buffer_.size();
     if (remaining == 0) {
